# docs/swagger.json
{
    "components": {
        "schemas": {
            "UserResponse": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "followers": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "followings": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "posts": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "bio": {
                        "type": "string"
                    },
                    "avatar": {
                        "type": "string"
                    },
                    "verified": {
                        "type": "boolean"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "username",
                    "email",
                    "firstName",
                    "lastName",
                    "followers",
                    "followings",
                    "posts",
                    "bio",
                    "avatar",
                    "verified"
                ]
            },
...
            "SignupRequest": {
                "properties": {
                    "username": {
                        "type": "string",
                        "maxLength": 20,
                        "minLength": 2
                    },
                    "password": {
                        "type": "string"
                    },
                    "email": {
                        "format": "email",
                        "type": "string"
                    },
                    "firstName": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 1
                    },
                    "lastName": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 1
                    }
                },
                "type": "object",
                "required": [
                    "username",
                    "password",
                    "email",
                    "firstName",
                    "lastName"
                ]
            },
            "LoginRequest": {
                "properties": {
                    "userIdentifier": {
                        "type": "string",
                        "minLength": 1
                    },
                    "password": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "type": "object",
                "required": [
                    "userIdentifier",
                    "password"
                ]
            },
            "PasswordResetRequest": {
                "properties": {
                    "token": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "type": "object",
                "required": [
                    "token",
                    "password"
                ]
            },
            "PasswordUpdateRequest": {
                "properties": {
                    "currentPassword": {
                        "type": "string",
                        "minLength": 1
                    },
                    "newPassword": {
                        "type": "string"
                    },
                    "terminateAllSessions": {
                        "type": "boolean"
                    }
                },
                "type": "object",
                "required": [
                    "currentPassword",
                    "newPassword"
                ]
            },
...
        }
    },
    "info": {
        "title": "SocialApp API",
        "version": "1.0.0",
        "description": "Generated with `routing-controllers-openapi`"
    },
    "openapi": "3.0.0",
    "paths": {
...
        "/auth/signup": {
            "post": {
                "operationId": "AuthController.signup",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SignupRequest"
                            }
                        }
                    },
                    "description": "SignupRequest",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "summary": "Sign up user",
                "tags": [
                    "Auth"
                ],
                "description": "\n    Registers a user with given credentials and sends a verification email that is valid for 24 hours. \n    It also auto-authenticates the user by generating and setting access and refresh tokens as cookies. \n    Unverified users can only access /auth/logout and /auth/user out of the protected endpoints.\n    Body parameters criteria are as follows:\n      Username: 2 to 20 characters, unique, and follow the specified format:\n        * Starts with a letter and ends with a letter or number.\n        * Contains only letters, numbers, hyphens, underscores, or periods.\n        * Does not contain consecutive hyphens, underscores, or periods.\n      Password: Minimum 8 characters, Should contain at least:\n        * One uppercase letter.\n        * One lowercase letter.\n        * One number.\n        * One special character.\n      Email: Must be a valid and unique email address.\n      First name: Up to 50 characters.\n      Last name: Up to 50 characters.\n    "
            }
        },
        "/auth/login": {
            "post": {
                "operationId": "AuthController.login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    },
                    "description": "LoginRequest",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "summary": "Authenticate user",
                "tags": [
                    "Auth"
                ],
                "description": "\n    Authenticates a user using either a username or email. This generates access and refresh tokens, set as cookies. \n    Users have limited access until email verification.  \n    Access tokens are verified on protected endpoint access. \n      Access token expires in 1 hour and refresh token expires in 24 hours.\n      Expired access tokens trigger refresh token rotation and new access token generation.\n      Inactivity for 24 hours triggers automatic logout.\n    "
            }
        },
        "/auth/logout": {
            "get": {
                "operationId": "AuthController.logout",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful response"
                    }
                },
                "summary": "Sign out user",
                "tags": [
                    "Auth"
                ],
                "description": "\n    Signs out a user by invalidating the access and refresh tokens.\n    Tokens are invalidated by adding them to the token denylist and clearing the cookies.\n    ",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/email/verify": {
            "put": {
                "operationId": "AuthController.verifyEmail",
                "parameters": [
                    {
                        "in": "query",
                        "name": "token",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful response"
                    }
                },
                "summary": "Verify email address",
                "tags": [
                    "Auth"
                ]
            }
        },
        "/auth/password": {
            "get": {
                "operationId": "AuthController.requestPasswordReset",
                "parameters": [
                    {
                        "in": "query",
                        "name": "email",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful response"
                    }
                },
                "summary": "Send password reset link",
                "tags": [
                    "Auth"
                ],
                "description": "\n    Sends a password reset link to the user's email address.\n    The link is valid for 6 hours.\n    "
            },
            "post": {
                "operationId": "AuthController.resetPassword",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PasswordResetRequest"
                            }
                        }
                    },
                    "description": "PasswordResetRequest",
                    "required": false
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful response"
                    }
                },
                "summary": "Reset user's password",
                "tags": [
                    "Auth"
                ],
                "description": "\n    Resets a user's password using the password reset token. All sessions are terminated after password reset.\n    "
            },
            "put": {
                "operationId": "AuthController.updatePassword",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PasswordUpdateRequest"
                            }
                        }
                    },
                    "description": "PasswordUpdateRequest",
                    "required": false
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful response"
                    }
                },
                "summary": "Update user's password",
                "tags": [
                    "Auth"
                ],
                "description": "\n    Updates a user's password and terminates all sessions if terminateAllSessions is true.\n    ",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/user": {
            "get": {
                "operationId": "AuthController.getAuthenticatedUser",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "summary": "Get authenticated user's info",
                "tags": [
                    "Auth"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    }
}